examples/healthcare_consultation/requirements.txt
# Requirements for healthcare consultation example
openai>=1.0.0
python-dotenv>=1.0.0
